pyproject.toml
[tool.poetry]
name = "flask-app-base"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.1"
flask_script = "^2.0.6"
flask_injector = "^0.12.0"
flask_cors = "^3.0.8"
python-dotenv = "^0.10.3"
flask_sqlalchemy = "^2.4.1"
psycopg2 = "^2.8.4"
flask-migrate = "^2.5.2"
flask-login = "^0.4.1"
flake8 = "^3.7.9"
faker = "^4.0.1"
ulid-py = "^0.0.12"
Flask-Session = "^0.3.1"
pyjwt = "^1.7.1"
werkzeug = "^0.16.0"
marshmallow = "^3.4.0"
marshmallow-sqlalchemy = "^0.22.0"
marshmallow_enum = "^1.5.1"
boto3 = "1.12.38"
gunicorn = "^20.0.4"
mysqlclient = "1.4.6"
sadisplay = "^0.4.9"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^4.3.21"
yapf = "^0.29.0"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
